Added: additional features and cleaned WebView2 implementation
Added a check to ensure the plugin only attempts to work with filetype ids containing LRM_INCONTEXT
Added a default pageview for any other filetype so users know what the plugin is for
Removed all the references and code using the referenced dlls for WebView2. Instead worked with the version installed to Trados Studio.